% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{CellRanker}
\alias{CellRanker}
\title{Performs full CellRanker Workflow}
\usage{
CellRanker(SeuratObj, vis = FALSE)
}
\arguments{
\item{SeuratObj}{Seurat Object Containing Counts. If Counts are processed, use
'CellRanker_noprocess' instead}

\item{vis}{Visualize Tsne Plot After Processing}
}
\description{
Performs full CellRanker Workflow
}
