% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{CellRanker_knownmarkers}
\alias{CellRanker_knownmarkers}
\title{Performs full CellRanker Workflow if marker genes are already known}
\usage{
CellRanker_knownmarkers(SeuratObj, marker.list, vis = FALSE)
}
\arguments{
\item{SeuratObj}{Seurat Object Containing Processed Counts.}

\item{marker.list}{A list of marker genes. Structured as a list of dataframes
that is the same length of the number of clusters at SeuratObj.meta.data$cell_type.
Each list entry contaings a dataframe with a single column of adjust pvalues labeled
"p_val_adj'. The rownames of the dataframe should be the respective gene. Only
Only intented for use in debugging purposes.}

\item{vis}{Visualize Tsne Plot After Processing}
}
\description{
Performs full CellRanker Workflow if marker genes are already known
}
