% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{celltype_ranking}
\alias{celltype_ranking}
\title{Facilitates the ranking of celltypes}
\usage{
celltype_ranking(celltype.list, pangao.list)
}
\arguments{
\item{celltype.list}{A list the same length as the number of clusters
that contains a vector of all possible celltyeps for each respective
cluster}

\item{pangao.listA}{list the same length as the number of clusters. Each list
entry ontains a sliced dataframe with information regarding each relevant marker
gene to the respective cluster.}
}
\description{
Facilitates the ranking of celltypes
}
