% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{cluster_assignments}
\alias{cluster_assignments}
\title{Determines the highest ranked cell type for each cluster. For duplicate clusters,
rankings are repeated with the next highest cell type until all cell type classifications
are unique.}
\usage{
cluster_assignments(celltype.list, ranking.list)
}
\arguments{
\item{ranking.list}{A list vectors containing numeric rankings for each cluster.}
}
\description{
#'#'@param celltype.list A list the same length as the number of clusters
that contains a vector of all possible celltyeps for each respective
cluster
}
