% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{find_markers}
\alias{find_markers}
\title{Finds the marker genes for each cluster}
\usage{
find_markers(Seuratobj)
}
\arguments{
\item{Seuratobj}{Seurat Object containing processed counts and cluster annotations}
}
\description{
Finds the marker genes for each cluster
}
