% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{get_cell_type_possibilities}
\alias{get_cell_type_possibilities}
\title{Returns all possible cell types for a given cluster based on marker genes}
\usage{
get_cell_type_possibilities(pangao.list)
}
\arguments{
\item{pangao.list}{A list the same length as the number of clusters. Each list
entry ontains a sliced dataframe with information regarding each relevant marker
gene to the respective cluster.}
}
\description{
Returns all possible cell types for a given cluster based on marker genes
}
