% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{get_initial_ranking}
\alias{get_initial_ranking}
\title{Helper function for celltype_ranking that intially ranks cell types
by number of occurances
#'@param celltype.list A list the same length as the number of clusters
that contains a vector of all possible celltyeps for each respective
cluster}
\usage{
get_initial_ranking(celltype.list, pangao.list)
}
\arguments{
\item{pangao.listA}{list the same length as the number of clusters. Each list
entry ontains a sliced dataframe with information regarding each relevant marker
gene to the respective cluster.}
}
\description{
Helper function for celltype_ranking that intially ranks cell types
by number of occurances
#'@param celltype.list A list the same length as the number of clusters
that contains a vector of all possible celltyeps for each respective
cluster
}
