% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{get_ref}
\alias{get_ref}
\title{Retreives dataset from Pangao database, which contains information about which
marker genes correspond to a given cell type}
\usage{
get_ref()
}
\description{
Retreives dataset from Pangao database, which contains information about which
marker genes correspond to a given cell type
}
