% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{label_seurat}
\alias{label_seurat}
\title{Adds final cell type annotations to Seurat Object}
\usage{
label_seurat(cluster.celltypes, Seuratobj)
}
\arguments{
\item{cluster.celltypes}{A vector containing cell type annotations for
each cluster.}

\item{Seuratobj}{A Seurat object with processed counts and tsne coordinates}
}
\description{
Adds final cell type annotations to Seurat Object
}
