% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{load_example_data}
\alias{load_example_data}
\title{Loads example prostate cancer dataset from Heidigger et. al}
\usage{
load_example_data()
}
\description{
Loads example prostate cancer dataset from Heidigger et. al
}
