% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{process_counts}
\alias{process_counts}
\title{Preprocesses counts by follwing Seurat's suggested preprocessing workflow}
\usage{
process_counts(Seuratobj)
}
\arguments{
\item{Seuratobj}{Seurat Object containing non-processed counts. Performs clustering}
}
\description{
Preprocesses counts by follwing Seurat's suggested preprocessing workflow
}
