% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{query_pangao}
\alias{query_pangao}
\title{Queries the pangao dataset for marker genes with signficant relations to
different cell types}
\usage{
query_pangao(marker.list)
}
\arguments{
\item{marker.list}{a list of dataframes containing marker genes as row names and
a column labeled p_val_adj that contains adusted pvalues for each gene. THe length
of the list corresponseds to the number of clusters}
}
\description{
Queries the pangao dataset for marker genes with signficant relations to
different cell types
}
