% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRanker Functions.R
\name{visualize}
\alias{visualize}
\title{Plots TSNE coordinates with appropriate CellRanker assigned labels}
\usage{
visualize(Seuratobj)
}
\arguments{
\item{Seuratobj}{a Seurat object with processed counts, tsne coordinates, and
CellRanker assigned labels}
}
\description{
Plots TSNE coordinates with appropriate CellRanker assigned labels
}
